updated overheat

+++ Assets/PublicVariableHandler.cs
@@ -30,8 +30,8 @@ public class PublicVariableHandler : MonoBehaviour
     public int playerLives;
     public int playerHealth;
     public float playerSpeed;
-    public int playerLaserHeatCap;
-    public int playerLaserHeatIncreaseAmount;
+    public float playerLaserHeatCap;
+    public float playerLaserHeatIncreaseAmount;
     public float laserHeatLossAfterSeconds;
 
     //Enemy Variables
